use OpenTelemetry class to wrap OTEL NodejsSDK and pino logger
